backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
types-requests==2.31.0.20240406
urllib3==2.2.1
